requirements.txt
Flask==2.0.2
finnhub-python==2.4.10
python-dateutil==2.8.2
